requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.2,<3.0.0
pydantic-settings==2.1.0

# LLM and AI
openai>=1.6.1,<2.0.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
litellm==1.17.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# Vector embeddings
sentence-transformers==2.2.2

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-extra-types==2.3.0
tenacity==8.2.3

# Safety and validation
jsonschema==4.20.0

# Monitoring and logging
loguru==0.7.2

# HTML parsing
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0

# PDF parsing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Telegram Bot
python-telegram-bot==20.7
